+++ lib/observableview.ts
@@ -6,7 +6,6 @@
 
 import {ViewNode, isComputingView} from './dnode';
 import SimpleEventEmitter from './simpleeventemitter';
-import {logLevel} from './core';
 import {IContextInfoStruct, Lambda} from './interfaces';
 import {deepEquals, once} from './utils';